+++ src/Hasklepias/Reexports.hs
@@ -16,17 +16,20 @@ module Hasklepias.Reexports (
     , module Control.Monad
     , module Control.Applicative
     , module Data.Bool
-    , module Data.Time.Calendar 
+    , module Data.Either
+    , module Data.Eq
     , module Data.Int
     , module Data.Maybe
     , module Data.Function
     , module Data.Functor
     , module Data.List
     , module Data.Ord
+    , module Data.Time.Calendar 
     , module Data.Text
     , module Data.Tuple
     , module Safe
     , module Flow
+    , module Witherable
 ) where
 
 import safe GHC.Num                         ( Integer )
@@ -36,11 +39,18 @@ import safe Data.Bool                       ( Bool(..)
                                             , (&&), not, (||)
                                             , bool
                                             , otherwise )
+import safe Data.Either                     ( Either(..))
+import safe Data.Eq                         ( Eq, (==))
 import safe Data.Function                   ( (.), ($), const, id )
 import safe Data.Functor                    ( fmap )
 import safe Data.Int                        ( Int )
-import safe Data.List                       ( all, map, filter, length, null
-                                            , zipWith )
+import safe Data.List                       ( all
+                                            , any
+                                            , map
+                                            , length
+                                            , null
+                                            , zipWith
+                                            , (++) )
 import safe Data.Maybe                      ( Maybe(..),
                                               maybe,
                                               isJust,
@@ -58,5 +68,6 @@ import safe Data.Time.Calendar              ( Day, MonthOfYear, Year
                                             , diffDays )
 import safe Data.Text                       ( pack, Text )
 import safe Data.Tuple                      ( fst, snd, uncurry, curry )
+import safe Witherable                      ( Filterable(filter) )
 import safe Flow                            ( (!>), (.>), (<!), (<.), (<|), (|>) )
 import Safe                                 ( headMay, lastMay )